Adding .setdoc .doc .rmdoc and updating .rp
This little addition is designed to work as follows: If a document has been set for the RP with .setdoc, then the bot will return the regular "The RP is..." but it will also include the document at the end, making it nice and clean imo. But, if a document __hasn't__ been added, it will act as normal. Feel free to add if you wish.